+++ src/firebase/functionCalls.js
@@ -18,4 +18,13 @@ export const initFunctionCalls = () => {
   functions.unsubmitTalk = buildFunctionWithTimezone('unsubmitTalk')
 }
 
+/**
+ * Cloud function can be long to start due to cold-start,
+ * so we call them once when connecting
+ */
+export const preloadFunctions = () => {
+  functions.submitTalk({ initialize: true })
+  functions.unsubmitTalk({ initialize: true })
+}
+
 export default functions