+++ src/layout/brand/brand.jsx
@@ -1,39 +1,55 @@
+/* eslint-disable jsx-a11y/anchor-is-valid */
 import React from 'react'
 import PropTypes from 'prop-types'
 import cn from 'classnames'
 import { Link } from 'redux-little-router'
 
+import OpenTrigger from 'components/helpers/openTrigger'
+import Portal from 'components/portals/portal'
+import withTheme from 'styles/themes/withTheme'
 import { withSizes } from 'styles/utils'
-import BrandMobile from './brandMobile'
-import './brand.css'
+
+import styles from './brand.module.css'
+
+const SidebarWrapper = withTheme(({ className, onClick, content }) => (
+  <Portal className={cn(className, styles.mobileSidebarPortal)}>
+    <div className={styles.mobileSidebar} onClick={onClick} role="button">
+      {content}
+    </div>
+  </Portal>
+))
 
 const Brand = ({
-  title, baseRoute, isMobile, hasSidebar, className,
-}) => {
-  if (isMobile && baseRoute !== '/') {
-    return (
-      <BrandMobile
-        className={cn('brand', className)}
-        hasSidebar={hasSidebar}
-      />
-    )
-  }
-  return (
-    <div className={cn('brand', className)}>
-      <Link href={baseRoute}>{title}</Link>
-    </div>
-  )
-}
+  title, baseRoute, isTablet, isMobile, sidebar, className,
+}) => (
+  <div className={cn(styles.brand, className)}>
+    {sidebar
+      && (isMobile || isTablet) && (
+        <OpenTrigger
+          renderTrigger={({ show, hide, isOpen }) => (
+            <a onClick={isOpen ? hide : show} role="button" className={styles.burgerLink}>
+              <i className={`fa ${isOpen ? 'fa-arrow-left' : 'fa-bars'}`} />
+            </a>
+          )}
+        >
+          {({ hide }) => <SidebarWrapper content={sidebar} onClick={hide} />}
+        </OpenTrigger>
+    )}
+    <Link href={baseRoute}>{title}</Link>
+  </div>
+)
 
 Brand.propTypes = {
   title: PropTypes.string.isRequired,
   isMobile: PropTypes.bool.isRequired,
-  hasSidebar: PropTypes.bool.isRequired,
+  isTablet: PropTypes.bool.isRequired,
   baseRoute: PropTypes.string.isRequired,
+  sidebar: PropTypes.node,
   className: PropTypes.string,
 }
 
 Brand.defaultProps = {
+  sidebar: undefined,
   className: undefined,
 }