+++ src/screens/organizer/events/events.jsx
@@ -2,18 +2,18 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { Link } from 'redux-little-router'
 
+import Badge from 'components/badge'
 import Titlebar from 'components/titlebar'
 import IconLabel from 'components/iconLabel'
 import Button from 'components/button'
 import { List, ListItem } from 'components/list'
-import RelativeDate from 'components/relativeDate'
+import EventDates from 'screens/components/eventDates'
 
-import { toDate } from 'helpers/firebase'
-import './events.css'
+import styles from './events.module.css'
 
 const MyEvents = ({ events, onSelect }) => (
-  <div className="events-page">
-    <Titlebar className="events-header" icon="fa fa-calendar-o" title="My events">
+  <div>
+    <Titlebar icon="fa fa-calendar-o" title="My events">
       <Button>
         {btn => (
           <Link href="/organizer/event/create" className={btn}>
@@ -23,14 +23,38 @@ const MyEvents = ({ events, onSelect }) => (
       </Button>
     </Titlebar>
     <List
-      className="events-content"
       array={events}
       noResult="No event yet !"
-      renderRow={({ id, name, updateTimestamp }) => (
+      renderRow={({
+        id, name, type, visibility, address, conferenceDates,
+      }) => (
         <ListItem
           key={id}
-          title={name}
-          subtitle={<RelativeDate date={toDate(updateTimestamp)} />}
+          title={<div className={styles.title}>{name}</div>}
+          subtitle={<IconLabel icon="fa fa-map-marker" label={address} />}
+          info={(
+            <div className={styles.infos}>
+              <div className={styles.badges}>
+                {visibility === 'private' && (
+                <Badge pill outline error={visibility === 'private'}>
+                  {visibility}
+                </Badge>
+                )}
+                <Badge
+                  pill
+                  outline
+                  success={type === 'meetup'}
+                  info={type === 'conference'}
+                  className={styles.type}
+                >
+                  {type}
+                </Badge>
+              </div>
+              {type === 'conference' && (
+              <EventDates dates={conferenceDates} className={styles.dates} />
+              )}
+            </div>
+          )}
           onSelect={() => onSelect(id)}
         />
       )}