+++ src/screens/speaker/sidebar/index.js
@@ -1,5 +1,1 @@
-import forRouter from 'hoc-little-router'
-import Sidebar from './sidebar.container'
-
-export { default as Sidebar } from './sidebar.container'
-export const SidebarMobile = forRouter('MOBILE_MENU')(Sidebar)
+export { default } from './sidebar.container'